jobqueue: Move jobqueue out of internal
