+++ internal/jobqueue/fsjobqueue/fsjobqueue.go
@@ -20,8 +20,8 @@ import (
 	"time"
 
 	"github.com/google/uuid"
+	"github.com/osbuild/osbuild-composer/pkg/jobqueue"
 
-	"github.com/osbuild/osbuild-composer/internal/jobqueue"
 	"github.com/osbuild/osbuild-composer/internal/jsondb"
 )
